Renames oauth to authDataManager in src
