+++ src/authDataManager/github.js
