+++ src/authDataManager/index.js
