+++ src/authDataManager/instagram.js
